pyproject.toml
[project]
name = "orin-pipeline"
version = "0.1.0"
description = "Song processing pipeline for the Orin music communication app"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Environment & Config
    "python-dotenv>=1.0.0",

    # Database
    # sqlite3 is built-in

    # Audio Processing
    "yt-dlp>=2024.0.0",

    # LLM APIs
    "groq>=0.4.0",
    "together>=1.0.0",

    # Embeddings
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",

    # Vector Database
    "qdrant-client>=1.7.0",

    # Cloud Storage (R2)
    "aioboto3>=12.0.0",

    # Utilities
    "tqdm>=4.65.0",
    "rich>=13.0.0",

    # API Server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sse-starlette>=2.0.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
orin-pipeline = "src.cli:run"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline/src"]